App load correct when screen rotate(BillFragment og MainActivity has been merged, so now only MainActivity exists). Search icon and layout functionallity is added, but it doesn't search in the BillList yet
