+++ app/src/main/java/asay/asaymobile/DatabaseTable.java
@@ -1,0 +1,9 @@
+package asay.asaymobile;
+
+/**
+ * Created by Andreas on 09-01-2018.
+ */
+
+public class DatabaseTable {
+    
+}
