+++ app/src/main/java/asay/asaymobile/activities/MainActivity.java
@@ -1,24 +1,53 @@
 package asay.asaymobile.activities;
 
+
+
+
+
+import android.app.SearchManager;
+import android.content.Context;
 import android.content.Intent;
+import android.graphics.Color;
+import android.graphics.PorterDuff;
 import android.os.Bundle;
+import android.support.design.widget.TabLayout;
+
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
+import android.support.v4.app.FragmentPagerAdapter;
+import android.support.v4.view.ViewPager;
 import android.support.v7.app.AppCompatActivity;
+
 import android.support.v7.widget.Toolbar;
+import android.view.Menu;
+import android.view.MenuInflater;
 import android.view.MenuItem;
+import android.widget.SearchView;
+
+import java.util.ArrayList;
 
 import asay.asaymobile.R;
-import asay.asaymobile.fragments.BillsFragment;
+import asay.asaymobile.fragments.BillsAllFragment;
+import asay.asaymobile.model.BillDTO;
 
 
 public class MainActivity extends AppCompatActivity {
     private Toolbar toolbar;
 
+    private SectionsPagerAdapter mSectionsPagerAdapter;
+
+
+    //Default user
+    private int userId = 1;
+    private boolean loggedIn = true;
+
+    private ViewPager mViewPager;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
+
 
         // Set a Toolbar to replace the ActionBar.
         toolbar = findViewById(R.id.toolbar);
@@ -27,19 +56,42 @@ public class MainActivity extends AppCompatActivity {
             getSupportActionBar().setHomeButtonEnabled(true);
         toolbar.setNavigationIcon(R.mipmap.ic_launcher);
 
-        // Instantiate fragment
-        Fragment fragment = null;
-        Class fragmentClass = BillsFragment.class;
-        try {
-            fragment = (Fragment) fragmentClass.newInstance();
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
 
-        // Insert the fragment by replacing any existing fragment
-        FragmentManager fragmentManager = getSupportFragmentManager();
-        fragmentManager.beginTransaction().replace(R.id.flContent, fragment).commit();
+
+
+
+
+
+        mSectionsPagerAdapter = new SectionsPagerAdapter(getSupportFragmentManager());
+        mViewPager = (ViewPager) findViewById(R.id.containerMain);
+        mViewPager.setAdapter(mSectionsPagerAdapter);
+
+        TabLayout tabLayout = (TabLayout) findViewById(R.id.tabsMain);
+        mViewPager.addOnPageChangeListener(new TabLayout.TabLayoutOnPageChangeListener(tabLayout));
+        tabLayout.addOnTabSelectedListener(new TabLayout.ViewPagerOnTabSelectedListener(mViewPager));
+        mViewPager.setCurrentItem(0,true);
+
+        tabLayout.getTabAt(0).getIcon().setColorFilter(Color.BLACK, PorterDuff.Mode.SRC_IN);
+        tabLayout.getTabAt(1).getIcon().setColorFilter(Color.parseColor("#a8a8a8"), PorterDuff.Mode.SRC_IN);
+        tabLayout.getTabAt(2).getIcon().setColorFilter(Color.parseColor("#a8a8a8"), PorterDuff.Mode.SRC_IN);
+        tabLayout.addOnTabSelectedListener(new TabLayout.ViewPagerOnTabSelectedListener(mViewPager){
+            //change the color on the selected tab icon to black
+            @Override
+            public void onTabSelected(TabLayout.Tab tab) {
+                tab.getIcon().setColorFilter(Color.BLACK, PorterDuff.Mode.SRC_IN);
+                super.onTabSelected(tab);
+            }
+            @Override
+            public void onTabUnselected(TabLayout.Tab tab) {
+                tab.getIcon().setColorFilter(Color.parseColor("#a8a8a8"), PorterDuff.Mode.SRC_IN);
+                super.onTabUnselected(tab);
+            }
+
+        });
+        mSectionsPagerAdapter.notifyDataSetChanged();
     }
+
+
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
@@ -52,4 +104,63 @@ public class MainActivity extends AppCompatActivity {
                 return super.onOptionsItemSelected(item);
         }
     }
+
+    public class SectionsPagerAdapter extends FragmentPagerAdapter {
+
+        public SectionsPagerAdapter(FragmentManager fm) {
+            super(fm);
+        }
+
+        @Override
+        public Fragment getItem(int position) {
+            Bundle bundle = new Bundle();
+            switch (position) {
+                case 0:
+                    BillsAllFragment billsAllFragment = new BillsAllFragment();
+                    return billsAllFragment;
+                case 1:
+                    bundle = new Bundle();
+                    bundle.putBoolean("isEnded",true);
+                    BillsAllFragment billsAllFragmentEnded = new BillsAllFragment();
+                    billsAllFragmentEnded.setArguments(bundle);
+                    return billsAllFragmentEnded;
+                case 2:
+                    if(loggedIn){
+                        bundle = new Bundle();
+                        bundle.putBoolean("isFavorite",true);
+                        BillsAllFragment billsAllFragmentFavorite = new BillsAllFragment();
+                        billsAllFragmentFavorite.setArguments(bundle);
+                        return billsAllFragmentFavorite;
+                    } else{
+                        return null;
+                    }
+
+                default:
+                    return null;
+            }
+        }
+
+        @Override
+        public int getCount() {
+            // Show 3 total pages.
+            return 3;
+        }
+
+    }
+
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        MenuInflater inflater = getMenuInflater();
+        inflater.inflate(R.menu.search_menu, menu);
+        // Associate searchable configuration with the SearchView
+        SearchManager searchManager =
+                (SearchManager) getSystemService(Context.SEARCH_SERVICE);
+        SearchView searchView =
+                (SearchView) menu.findItem(R.id.search).getActionView();
+        searchView.setSearchableInfo(
+                searchManager.getSearchableInfo(getComponentName()));
+        return true;
+    }
+
+
 }
