--- app/src/main/java/asay/asaymobile/fragments/BillsFragment.java
@@ -1,131 +1,0 @@
-package asay.asaymobile.fragments;
-
-import android.graphics.Color;
-import android.graphics.PorterDuff;
-import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.design.widget.TabLayout;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentPagerAdapter;
-import android.support.v4.view.ViewPager;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import asay.asaymobile.R;
-
-
-public class BillsFragment extends Fragment{
-
-    /**
-     * The {@link android.support.v4.view.PagerAdapter} that will provide
-     * fragments for each of the sections. We use a
-     * {@link FragmentPagerAdapter} derivative, which will keep every
-     * loaded fragment in memory. If this becomes too memory intensive, it
-     * may be best to switch to a
-     * {@link android.support.v4.app.FragmentStatePagerAdapter}.
-     */
-    private BillsFragment.SectionsPagerAdapter mSectionsPagerAdapter;
-    private int userId = 1;
-    private boolean loggedIn = true;
-    private Bundle bundle;
-    /**
-     * The {@link ViewPager} that will host the section contents.
-     */
-    private ViewPager mViewPager;
-
-    public BillsFragment() {
-        // Required empty public constructor
-    }
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        View rootView = inflater.inflate(R.layout.fragment_bills,container,false);
-        return rootView;
-    }
-    @Override
-    public void onViewCreated(final View rootView, @Nullable Bundle savedInstanceState) {
-        super.onViewCreated(rootView, savedInstanceState);
-
-
-        // Create the adapter that will return a fragment for each of the three
-        // primary sections of the activity.
-        final TabLayout tabLayout = getActivity().findViewById(R.id.tabs);
-        mSectionsPagerAdapter = new SectionsPagerAdapter(getFragmentManager());
-
-        // Set up the ViewPager with the sections adapter.
-        mViewPager = getActivity().findViewById(R.id.container);
-        mViewPager.setAdapter(mSectionsPagerAdapter);
-
-        mViewPager.addOnPageChangeListener(new TabLayout.TabLayoutOnPageChangeListener(tabLayout));
-
-        //set color filter to grey
-        tabLayout.getTabAt(0).getIcon().setColorFilter(Color.BLACK, PorterDuff.Mode.SRC_IN);
-        tabLayout.getTabAt(1).getIcon().setColorFilter(Color.parseColor("#a8a8a8"), PorterDuff.Mode.SRC_IN);
-        tabLayout.getTabAt(2).getIcon().setColorFilter(Color.parseColor("#a8a8a8"), PorterDuff.Mode.SRC_IN);
-        tabLayout.addOnTabSelectedListener(new TabLayout.ViewPagerOnTabSelectedListener(mViewPager){
-            //change the color on the selected tab icon to black
-            @Override
-            public void onTabSelected(TabLayout.Tab tab) {
-                tab.getIcon().setColorFilter(Color.BLACK, PorterDuff.Mode.SRC_IN);
-                super.onTabSelected(tab);
-            }
-            @Override
-            public void onTabUnselected(TabLayout.Tab tab) {
-                tab.getIcon().setColorFilter(Color.parseColor("#a8a8a8"), PorterDuff.Mode.SRC_IN);
-                super.onTabUnselected(tab);
-            }
-
-        });
-        mSectionsPagerAdapter.notifyDataSetChanged();
-    }
-
-
-    /**
-     * A {@link FragmentPagerAdapter} that returns a fragment corresponding to
-     * one of the sections/tabs/pages.
-     */
-    public class SectionsPagerAdapter extends FragmentPagerAdapter {
-
-        public SectionsPagerAdapter(FragmentManager fm) {
-            super(fm);
-        }
-
-        @Override
-        public Fragment getItem(int position) {
-            switch (position) {
-                case 0:
-                    BillsAllFragment billsAllFragmentActive = new BillsAllFragment();
-                    return billsAllFragmentActive;
-                case 1:
-                    bundle = new Bundle();
-                    bundle.putBoolean("isEnded",true);
-                    BillsAllFragment billsAllFragmentEnded = new BillsAllFragment();
-                    billsAllFragmentEnded.setArguments(bundle);
-                    return billsAllFragmentEnded;
-                case 2:
-                    if(loggedIn){
-                        bundle = new Bundle();
-                        bundle.putBoolean("isFavorite",true);
-                        BillsAllFragment billsAllFragmentFavorite = new BillsAllFragment();
-                        billsAllFragmentFavorite.setArguments(bundle);
-                        return billsAllFragmentFavorite;
-                    } else{
-                        return null;
-                    }
-
-                default:
-                    return null;
-            }
-        }
-
-        @Override
-        public int getCount() {
-            // Show 3 total pages.
-            return 3;
-        }
-    }
-
-}
