+++ app/src/main/res/values/strings.xml
@@ -37,5 +37,8 @@
     <string name="your_profile">Din profil</string>
     <string name="error_no_text">Du har ikke indtastet tekst i feltet</string>
     <string name="noDesc">Der findes ingen beskrivelse til dette forslag</string>
+    <string name="active">Aktive</string>
+    <string name="Ended">Afsluttede</string>
+    <string name="favorits">Favoritter</string>
 
 </resources>
