+++ app/src/main/res/xml/searchable.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<searchable xmlns:android="http://schemas.android.com/apk/res/android"
+    android:label="Søg her"
+    android:hint="søg">
+    </searchable>
